+++ modules/img2img.py
@@ -10,7 +10,7 @@ from modules.memstats import memory_stats
 
 
 def process_batch(p, input_files, input_dir, output_dir, inpaint_mask_dir, args):
-    shared.log.debug(f'batch: {input_dir}|{output_dir}|{inpaint_mask_dir}')
+    shared.log.debug(f'batch: {input_files}|{input_dir}|{output_dir}|{inpaint_mask_dir}')
     processing.fix_seed(p)
     if input_files is not None and len(input_files) > 0:
         image_files = [f.name for f in input_files]
@@ -109,8 +109,11 @@ def img2img(id_task: str, mode: int, prompt: str, negative_prompt: str, prompt_s
 
     shared.log.debug(f'img2img: id_task={id_task}|mode={mode}|prompt={prompt}|negative_prompt={negative_prompt}|prompt_styles={prompt_styles}|init_img={init_img}|sketch={sketch}|init_img_with_mask={init_img_with_mask}|inpaint_color_sketch={inpaint_color_sketch}|inpaint_color_sketch_orig={inpaint_color_sketch_orig}|init_img_inpaint={init_img_inpaint}|init_mask_inpaint={init_mask_inpaint}|steps={steps}|sampler_index={sampler_index}|latent_index={latent_index}|mask_blur={mask_blur}|mask_alpha={mask_alpha}|inpainting_fill={inpainting_fill}|full_quality={full_quality}|restore_faces={restore_faces}|tiling={tiling}|n_iter={n_iter}|batch_size={batch_size}|cfg_scale={cfg_scale}|image_cfg_scale={image_cfg_scale}|clip_skip={clip_skip}|denoising_strength={denoising_strength}|seed={seed}|subseed{subseed}|subseed_strength={subseed_strength}|seed_resize_from_h={seed_resize_from_h}|seed_resize_from_w={seed_resize_from_w}|selected_scale_tab={selected_scale_tab}|height={height}|width={width}|scale_by={scale_by}|resize_mode={resize_mode}|inpaint_full_res={inpaint_full_res}|inpaint_full_res_padding={inpaint_full_res_padding}|inpainting_mask_invert={inpainting_mask_invert}|img2img_batch_files={img2img_batch_files}|img2img_batch_input_dir={img2img_batch_input_dir}|img2img_batch_output_dir={img2img_batch_output_dir}|img2img_batch_inpaint_mask_dir={img2img_batch_inpaint_mask_dir}|override_settings_texts={override_settings_texts}')
 
-    if init_img is None:
-        shared.log.debug('Init image not set')
+    if mode == 5:
+        if img2img_batch_files is None or len(img2img_batch_files) == 0:
+            shared.log.debug('Init bactch images not set')
+        elif init_img:
+            shared.log.debug('Init image not set')
 
     if sampler_index is None:
         sampler_index = 0
@@ -202,7 +205,6 @@ def img2img(id_task: str, mode: int, prompt: str, negative_prompt: str, prompt_s
         inpainting_mask_invert=inpainting_mask_invert,
         override_settings=override_settings,
     )
-    p.is_batch = mode == 5
     if selected_scale_tab == 1 and resize_mode != 0:
         p.scale_by = scale_by
     p.scripts = modules.scripts.scripts_img2img
@@ -210,6 +212,7 @@ def img2img(id_task: str, mode: int, prompt: str, negative_prompt: str, prompt_s
     p.extra_generation_params['Resize mode'] = resize_mode
     if mask:
         p.extra_generation_params["Mask blur"] = mask_blur
+    p.is_batch = mode == 5
     if p.is_batch:
         process_batch(p, img2img_batch_files, img2img_batch_input_dir, img2img_batch_output_dir, img2img_batch_inpaint_mask_dir, args)
         processed = processing.Processed(p, [], p.seed, "")